add a hack to hide entries from the compdb
As explained in the comments in this commit, each compdb entry has a file key to indicate which input file the command applies to. `ninja` uses the first file in $in as the "file" key for the compdb files it generates which is problematic for us because we have format targets that take the source files as inputs ($in). Tools (at least `clang-format` and friends) that use compdb don't deduplicate: if there are multiple commands for a particular file it'll use all those commands (for `clang-format` it'll go try to format the file multiple times, I think?). The entries for source files that have format commands and the like understandably confuse `clang-format` (it's expecting arguments to pass to `clang` so it doesn't understand the flags/arguments).
The way we handle this is a little gross:
  - We have a designated $unused_file (that ends in `.c` for a reason explained below).
  - For our format commands we pass this file in first. That way their comamnd entries have $unused_file as their "file" key.
  - We also go create $unused_file (in common.ninja!! this is the only build rule in common.ninja) so that the file actually exists so ninja doesn't get mad.
  - This alone isn't enough. We've now got all the commands that we don't actually want in the compdb to have $unused_file as their "file" key, which is good. But:
    + in order to do that we had to go pass in $unused_file to those commands
    + those commands read the compdb
    + and we just told those commands to operate on $unused_file
    + meaning they'll still see their own commands in compdb and get confused!
  - To close the loop, we also have to go purge the entries with $unused_file from the compdb, which is the second part of this commit (more details below).

Tools also seem to be very insistent that all compdb entries have all the necessary fields.
As mentioned in a comment in `common.ninja` in this commit (this will disappear from the source in the next cleanup commit so I'm putting it in the commit message too for record keeping):

`'[.[] | select(.file != "$builddir/$unused_file")] + [{"file": "$builddir/$unused_file"}]' > $out` (i.e. adding a basically fake entry with the unused file so tools aren't confused when they can't find the file in the compdb) doesn't work.
this is super weird: clang-tidy will warn about $unused_file not having a compile command if we put in a compdb entry for it but have an empty compile command. But if we just omit the entry altogether, it happily chugs along (no warnings, no errors).

Also note: if we don't include directory and compile keys, it'll just die (fails to use the compdb at all).
Also bizarre: if the file ($unused_file) doesn't end in an extension that clang-tidy recognizes, it'll hard error (so we're using .c for now).