update to llvm-9!
The lldb-9 package (for xenial at least) is wonky and as a result the Dockerfile is more closely coupled to the LLVM version and the Ubuntu version than I'd like.
Also, clangd is it's own package now.

Other than Dockerfile changes we also get to swap some of the flags on the clang tools to use two hypens instead of one _as they should_. This worked fine earlier too, but now that the docs have changed too we can switch without confusing anyone.