Switch to a fork of ninja
The fork includes ninja-build#1578 (with a change) which we want for modules.

The old compdb patch has made it into master (but not yet an offical release) so we no longer have to apply it manually.
(https://github.com/ninja-build/ninja/commit/20af31d586c1f1bddeaad0583adc17933d5465ce)